+++ ontimize/components/input/combo/o-combo.component.ts
@@ -114,9 +114,9 @@ export class OComboComponent extends OFormServiceComponent implements OnInit, Af
     return this.nullSelection;
   }
 
-  syncDataIndex() {
-    super.syncDataIndex();
-    if (this.nullSelection && Util.isDefined(this._currentIndex)) {
+  syncDataIndex(queryIfNotFound: boolean = true) {
+    super.syncDataIndex(queryIfNotFound);
+    if (this._currentIndex !== undefined && this.nullSelection) {
       // first position is for null selection that it is not included into dataArray
       this._currentIndex += 1;
     }